requirements.txt
pip
setuptools
wheel

python-dotenv
requests

Flask

SQLAlchemy

bcrypt

pyjwt